+++ src/mongodb_session_manager/__init__.py
@@ -124,6 +124,6 @@ __all__.extend(
     ]
 )
 
-__version__ = "0.10.0"
+__version__ = "0.10.1"
 __author__ = "Iñaki Guinea Beristain"
 __author_email__ = "iguinea@gmail.com"